Make Gherkin Scanner Feature Complete
* Successfully tokenizes all feature files as
documented [here](https://github.com/cucumber/cucumber/tree/master/gherkin/testdata/good)

On the one hand, in order to benefit from the wealth of the testing data
referred to above, `test/support/scanner_support.ex` takes the output
from `Scanner` and converts the same to the `.tokens` format which is
then compared to the `.tokens` files provided therein. Passing all those
cases should provide enough confidence that all features as well as all
major edge cases have been met.

On the other hand, `Scanner` is not intended to adhere to the `.tokens`
format as `Erlang` has good support for `yecc` parser which requires a
different format altogether.

* Full LTR Support of [all documented languages](https://github.com/cucumber/cucumber/blob/master/gherkin/gherkin-languages.json)

Part of the test cases referred to above did use different languages,
such as French, Norwegian and even the emoji language. However, RTL
examples and test cases were rather absent so will need to be addressed
later on, which will require a variety of example files.

* A Mix Task that can take `gherkin-languages.json` as input and
convert the same in a slightly better format. The idea here is that the
current format does not denote directly if a given language is `LTR` or
`RTL`. While this can be derived from the name property with a `Regex`
scan runtime, or even compile time; to have this information directly
embedded is the better way forward.