Introduce Language Support Config Options
The main purpose for this functionality is to
reduce the long compile time on part of scanner.ex
that needs to consume the full version of 'gherkin-
language.json'.

As a feature, this should allow consumers of this
library to create their own set of Gherkin keywords
that may make more sense in their Domain of concern.