Introduce complex feature files for testing
The main purpose of this commit is to ensure that `Gherkin.Scanner`
behaves properly when confronted with complex, fully-fledged feature
files.

A great multitude of feature files have been downloaded from the project:
`cucumber-ruby` and even some blog posts were scraped to ensure the
richness required to conform to the Gherkin spec as is only known by the
official Gherkin scanners.

The online documentation available at `cucumber.io` only covers a basic
subset of the actual specification. It is for exactly this reason that
the help of a `ruby` script: `generate-tokens`(which is based on
`cucumber-ruby`) was invoked so that the desired baseline for all those
'.feature'-files could be met. `Gherkin.Scanner` could then accordingly
be fully tested.