Introduce Feature Complete Gherkin Parser
As inspired by `Elixir` itself; this parser is based on `yecc` which you
may find under `src/parser.yrl`. The `grammar` detailed therein is
effectively a translation of the `grammar` as outlined by the main
`Cucumber`-project which you may find [here](
https://github.com/cucumber/cucumber/blob/master/gherkin/gherkin.berp)

This parser fulfills two major responsibilities:
* Accepting `feature`-files that adhere to the `Gherkin`-grammar and
converting the same to a traversable `AST`-format.
* Rejecting `feature`-files that contradict said grammar by throwing a
syntax error. It is the aim of this project to make these very clear and
informative for the end-user. For this purpose; a proof of concept has
been coined under the abstraction of `syntax_error_message` to embed a
hyperlink that allows to discuss the error at hand in greater length
than is practical under the context of a terminal console.

Producing `AST` is only beneficial when the same can be traversed. To
facilitate this, `ast_traverser` has been introduced that allows the
end-user to fully traverse the tree with a set of functions to
aggregate à volonté. This same traverser should come in handy for the
`Cucumber`-library. The idea was to create a generic `AST`-traverser
that is fully customizable. Regretfully, the current implementation is
highly complex. Unless one is intimately aware with the underlying
implementation; the advice is to rather write one's own customized
solution until a clearer API is provided.

In order to proof that this parser works, this project was ambitious
enough to translate the `AST` produced by this parser to `ast.ndjson`-
format from the `Cucumber`-project. Like `Tokenizer`, the tests compare
the output generated with the `ast.ndjson` files. Converting `AST` to
`ast.ndjson`-format was done using `AST`-traverser discussed above.