Introduce Feature Complete Gherkin Parser
The parser introduced is based on `yecc` parser for which erlang provides
first class citizen support. Besides this convenience, the main reason for
going this route was inspired by the fact that `Elixir` too uses a `yecc`
parser to produce an `AST`; which is the main purpose of the parser. In
addition to this functionality, a parser also is responsible for rejecting
that which does not conform to the grammar at hand in the form of syntax
errors. In this regard, the ambition of this project is to make these
syntax errors very detailed and descriptive so that the end user of this
grammar is not left frustrated with resolving ambiguity. A proof of
concept has already been made by introducing `syntax_error_message` as
an abstraction to encapsulate a variety of concerns. This includes a link
to a webpage that is intended to discuss the error at hand in greater
length than is practical in a terminal console.
Producing an `AST` is only beneficial when the same can be traversed. To
facilitate this, `ast_traverser` has been introduced that allows a
library user to fully traverse the tree with a set of functions to
aggregate a la volente. This same traverser should come in handy for the
`Cucumber` library. The idea was to create a generic tree AST traverser
that is fully customizable. Regretfully, the current implementation is
highly complex and unless one is intimately aware with the underlying
implementation; the advice is to rather write your own customized
solution until a clearer API is provided.
In order to proof that this parser works, this project was ambitious
enough to translate the `AST` produced by this parser to `ast.ndjson`-
format from the Cucumber project. Like the tokenizer, the tests compare
the output generated with the `ast.ndjson` files. Converting the `AST` to
`ast.ndjson` format was done using the `ast_traverser` discussed above.